Suppress some warnings about size_t -> int conversions

+++ cxx4/ncCompoundType.cpp
@@ -88,7 +88,7 @@ void NcCompoundType::addMember(const string& memberName, const NcType& newMember
 void NcCompoundType::addMember(const string& memberName, const NcType& newMemberType, size_t offset, const vector<int>& shape)
 {
   if (!shape.empty())
-    ncCheck(nc_insert_array_compound(groupId, myId,const_cast<char*>(memberName.c_str()), offset, newMemberType.getId(), shape.size(), const_cast<int*>(&shape[0])),__FILE__,__LINE__);
+    ncCheck(nc_insert_array_compound(groupId, myId,const_cast<char*>(memberName.c_str()), offset, newMemberType.getId(), static_cast<int>(shape.size()), const_cast<int*>(&shape[0])),__FILE__,__LINE__);
   else
     addMember(memberName, newMemberType, offset);
 }
